% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxminmin.R
\name{max_min_min}
\alias{max_min_min}
\title{Max - Min}
\usage{
max_min_min(x)
}
\arguments{
\item{x}{numeric}
}
\value{
numeric
}
\description{
Combines baseR functions of max - min. Later will build into another, more generalized function...
}
\examples{
#' max_min_min(iris$Sepal.Length), max_min_min(iris$Petal.Length)
}

