% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxmin-package.r
\docType{package}
\name{maxmin}
\alias{maxmin}
\alias{maxmin-package}
\title{maxmin: find the diff in quantiles}
\description{
A crazy simple use of baseR max and min to find the difference. Then a slightly more interesting generalized version that can take any quantile of interest to the user.
qdiff includes a couple different arguments including na.rm and control of which type of algorithm generates the quantile.
}

