% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdiff.R
\name{qdiff}
\alias{qdiff}
\title{Quantile - Quantile}
\usage{
qdiff(x, probs = c(0, 1), na.rm = TRUE, ...)
}
\arguments{
\item{x}{numeric}

\item{probs}{the quantiles of interest}

\item{na.rm}{TRUE or FALSE: what do you want to do with NAs?}

\item{...}{9 different algorithms for computing quantiles}
}
\value{
numeric
}
\description{
Quantile - Quantile
}
\examples{
#' set.seed(1234)
#' z <- rnorm(10)
#' qdiff(z, probs = c(0.25, 0.75), type = 1)
}

